add pre-commit hook and update README

+++ core/extractors/clip/clip_extractors.py
@@ -10,11 +10,10 @@ from .utils import CLIP
 
 from .. import EXTRCT_REGISTRY
 
+
 class ClipExtractorBase(nn.Module):
-    def __init__(
-        self, 
-        model_name: str = 'ViT-B/32',
-        **kwargs) -> None:
+
+    def __init__(self, model_name: str = 'ViT-B/32', **kwargs) -> None:
         super().__init__()
 
         clip_state_dict = CLIP.get_config(model_name=model_name)
@@ -26,10 +25,13 @@ class ClipExtractorBase(nn.Module):
 
         # set the parameters of CLIP
         vision_width = clip_state_dict["visual.conv1.weight"].shape[0]
-        vision_layers = len(
-            [k for k in clip_state_dict.keys() if k.startswith("visual.") and k.endswith(".attn.in_proj_weight")])
+        vision_layers = len([
+            k for k in clip_state_dict.keys()
+            if k.startswith("visual.") and k.endswith(".attn.in_proj_weight")
+        ])
         vision_patch_size = clip_state_dict["visual.conv1.weight"].shape[-1]
-        grid_size = round((clip_state_dict["visual.positional_embedding"].shape[0] - 1) ** 0.5)
+        grid_size = round(
+            (clip_state_dict["visual.positional_embedding"].shape[0] - 1)**0.5)
         image_resolution = vision_patch_size * grid_size
 
         embed_dim = clip_state_dict["text_projection"].shape[1]
@@ -37,22 +39,33 @@ class ClipExtractorBase(nn.Module):
         vocab_size = 49408
         transformer_width = clip_state_dict["ln_final.weight"].shape[0]
         transformer_heads = transformer_width // 64
-        transformer_layers = len(set(k.split(".")[2] for k in clip_state_dict if k.startswith(f"transformer.resblocks")))
+        transformer_layers = len(
+            set(
+                k.split(".")[2] for k in clip_state_dict
+                if k.startswith(f"transformer.resblocks")))
 
         cut_top_layer = 0
-        
+
         model = CLIP(
             embed_dim,
-            image_resolution, vision_layers-cut_top_layer, vision_width, vision_patch_size,
-            context_length, vocab_size, transformer_width, transformer_heads, transformer_layers-cut_top_layer,
+            image_resolution,
+            vision_layers - cut_top_layer,
+            vision_width,
+            vision_patch_size,
+            context_length,
+            vocab_size,
+            transformer_width,
+            transformer_heads,
+            transformer_layers - cut_top_layer,
         ).float()
 
         ret = model.load_state_dict(clip_state_dict, strict=False)
         return model, embed_dim
 
+
 @EXTRCT_REGISTRY.register()
 class ClipVideoExtractor(ClipExtractorBase):
-    
+
     def __init__(self, model_name: str = 'ViT-B/32', **kwargs) -> None:
         super().__init__(model_name=model_name)
         self.model = self.clip.visual
@@ -96,14 +109,15 @@ class ClipVideoExtractor(ClipExtractorBase):
 
         return pooled_output
 
+
 @EXTRCT_REGISTRY.register()
 class ClipImageExtractor(ClipExtractorBase):
-    
+
     def __init__(self, model_name: str = 'ViT-B/32', **kwargs) -> None:
         super().__init__(model_name=model_name)
         self.model = self.clip.visual
         self.clip = None
-        
+
     def forward(self, x):
         """video encoder
         Returns:
@@ -114,9 +128,10 @@ class ClipImageExtractor(ClipExtractorBase):
         x = hidden[:, 0, :]
         return x
 
+
 @EXTRCT_REGISTRY.register()
 class ClipTextExtractor(ClipExtractorBase):
-    
+
     def __init__(self, model_name: str = 'ViT-B/32', **kwargs) -> None:
         super().__init__(model_name=model_name)
         self.model = self.clip.visual
@@ -135,7 +150,8 @@ class ClipTextExtractor(ClipExtractorBase):
         Returns:
             x: output embedding [1,512]
         """
-        x = self.token_embedding(batch["input_ids"]).type(self.dtype)  # [batch_size, n_ctx, d_model]
+        x = self.token_embedding(batch["input_ids"]).type(
+            self.dtype)  # [batch_size, n_ctx, d_model]
 
         pos_emd = self.positional_embedding[:x.size(1), :].type(self.dtype)
         x = x + pos_emd
@@ -147,6 +163,7 @@ class ClipTextExtractor(ClipExtractorBase):
 
         # x.shape = [batch_size, n_ctx, transformer.width]
         # take features from the eot embedding (eot_token is the highest number in each sequence)
-        x = hidden[torch.arange(hidden.shape[0]), batch["input_ids"].argmax(dim=-1)]
+        x = hidden[torch.arange(hidden.shape[0]),
+                   batch["input_ids"].argmax(dim=-1)]
 
         return x
